QuantumSpinInverter Cost reduction
The quantum spin inverter in it's current state costs far too many resources for the effect it has. Admittedly, when you make it the one use teleport is VERY powerful. However, considering that epistemics tends to not be provided too many resources by cargo and usually it's members don't care about moving since they are usually just sitting in epistemics, it disproportionally benefits syndies and those with cargo access who can actually get the resources needed to make a bunch of the things. A cost reduction would allow more people to use it than just syndies and also make using it less of a commitment. I myself find that i use them as a emergency TP only. Being cheaper would allow people to use it for more frivolous uses such as on box teleporting straight from cargo to epistemics. Furthermore, it would provide epistemics the capability to distribute the things to others outside of the department due to decreased cost.