Add components to prototypes to help tests pass.
Two tests fail when I run the entire test suite (TestRestockBreaksOpen
and TestCompleteRestockProcess) but none of the tests fail when I run it
with --filter VendingMachineRestock. I can only assume that other tests
are somehow interfering due to the mixed environment of how tests run
one after another. This commit should only be an issue if the
natural parents lose Wires or Damageable.