Fix markings being not completely visible
Took the easy way out of this issue by just adding the missing chest markings on the legs/arm markings. It's dumb and pretty much not the proper way to fix it but it'll do until it layering isn't dumb either.
Also fixed hands and foot not getting markings due to them appearing above the legs.