Rename and fix LoadTickLoadBagel test
bagel.yml on upstream has the SpreaderGrid's nextUpdate set to 1 and
this apparently keeps LoadTickLoadBagel from failing the test.

No other upstream map at this time of writing has their SpreaderGrid
set to anything other than 0 or unset.

This commit also renames the test to LoadTickLoadAngle to reflect the
map we use for testing.