Fix lack of gravity at round-start
This commit introduces a patch to GravityGeneratorComponent which allows
GravityActive to be set to true in a map file.

Gravity generators are set to have a full charge and be switched on at
init, but they do not produce gravity on init. This was half of the
cause for the lack of round-start gravity. The other half is that the
grid's Gravity component is not set to true.

Setting one side is not enough. Both must be true in order to avoid
firing the GravityChangedEvent.

Gravity generators were left as GravityActive = false on init
in order to comply with the expectations of tests.