Shipwrecked patch 02 (#1684)
* Lower max player count to 7

* Adjust spawn chances of shipwreck factions

* Adjust shipwreck destination structure count

* Make destination temperatures more extreme

* Update ruined shops dungeon map

Adds new shop: Hippies

Removes the gun safe in the gun shop in favor of just pre-spawned ammo
and shotgun. StorageFill on the safe was not functioning.

* Start with full SMES for ruined buildings

* Expand random structure spot distance

* Tweak how the generator objective is determined

It now functions off of the initial load of the shuttle, rather than the
maximum supply of the generators. This should make things a little
easier on the players.

* Stop burning thrusters when they decouple

* Cleanup a section of code

* Damage the generator(s) more intelligently

* Update changelog