Add Blue Curacao Recipe of OJ and Sugar (#192)
* Add Blue Curacao Recipe of OJ and Sugar

'Curaçao is a Caribbean liqueur made using the dried peel of the Laraha citrus fruit.' We do not have the Laraha citrus fruit, but we have OJ and artificial food sweetener in the form of sugar, and I think that's close enough.

* Fixed the file placement.

Removed from upstream drinks list. Added to Delta-V drinks list.